AutoSummarizer/requirements.txt
fastapi
uvicorn
pymupdf
python-dotenv
google-generativeai
langchain
langchain-chroma
langchain-google-genai
camelot-py
opencv-python
pillow
numpy
